Extract Auction fixture for better code reusability
